remove UTXO query from Bitcoin RPC status check, we have utxo metrics already
