+++ zetaclient/chains/ton/liteapi/client_live_test.go
@@ -75,7 +75,7 @@ func TestClient(t *testing.T) {
 		})
 	})
 
-	t.Run("GetTransactionsUntil", func(t *testing.T) {
+	t.Run("GetTransactionsSince", func(t *testing.T) {
 		// ARRANGE
 		// Given sample account id (dev wallet)
 		// https://tonviewer.com/UQCVlMcZ7EyV9maDsvscoLCd5KQfb7CHukyNJluWpMzlD0vr?section=transactions
@@ -166,6 +166,18 @@ func TestClient(t *testing.T) {
 		assert.LessOrEqual(t, since, 20*time.Second)
 
 		t.Logf("Masterchain block #%d is generated at %q (%s ago)", block.SeqNo, blockTime, since.String())
+	})
+
+	t.Run("HealthCheck", func(t *testing.T) {
+		// ACT
+		blockTime, err := client.HealthCheck(ctx)
+
+		// ASSERT
+		require.NoError(t, err)
+
+		// Check that block was generated less than 20 seconds ago
+		since := time.Since(blockTime)
+		require.LessOrEqual(t, since, 20*time.Second)
 	})
 
 	t.Run("GetGasConfig", func(t *testing.T) {
